+++ hutool-core/src/main/java/cn/hutool/core/io/FileUtil.java
@@ -3393,6 +3393,20 @@ public class FileUtil {
 	public static boolean isSymlink(File file) throws IORuntimeException {
 		return Files.isSymbolicLink(file.toPath());
 	}
+	
+	/**
+	 * 判断给定的目录是否为给定文件或文件夹的父目录
+	 * 
+	 * @param parent 父目录
+	 * @param sub 子目录
+	 * @return 子目录是否为父目录的子目录
+	 * @since 4.5.4
+	 */
+	public static boolean isSub(File parent, File sub) {
+		Assert.notNull(parent);
+		Assert.notNull(sub);
+		return sub.toPath().startsWith(parent.toPath());
+	}
 
 	/**
 	 * 创建{@link RandomAccessFile}